ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Google Cloud Vision OCR
google-cloud-vision==3.7.2

# Google Gemini
google-generativeai==0.5.4

# Face recognition / embeddings
deepface==0.0.93
tensorflow==2.16.1
opencv-python-headless==4.9.0.80

# Structured logging
python-json-logger==2.0.7

# Utility
pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1
